refactor: reorganize Sanity studio schemas and add custom structure
- Move document schemas to schemaTypes/documents/
- Move singleton schemas to schemaTypes/singletons/
- Export singletonTypes set for filtering "New document" menu
- Add custom structure: Content group (blog, experience, skills) + Site Settings singleton
- siteSettings links directly to a fixed documentId - no list, no duplicates
- Remove __experimental_actions (structure builder handles singleton enforcement)
