+++ src/index.css
@@ -71,7 +71,8 @@
   text-decoration: underline;
 }
 
-.blog-content ul, .blog-content ol {
+.blog-content ul,
+.blog-content ol {
   padding-left: 1.5rem;
   margin-bottom: 1rem;
 }
