+++ studio/sanity.config.ts
@@ -1,7 +1,31 @@
 import {defineConfig} from 'sanity'
 import {structureTool} from 'sanity/structure'
 import {visionTool} from '@sanity/vision'
-import {schemaTypes} from './schemaTypes'
+import {schemaTypes, singletonTypes} from './schemaTypes'
+import type {StructureResolver} from 'sanity/structure'
+
+const structure: StructureResolver = (S) =>
+  S.list()
+    .title('Portfolio')
+    .items([
+      S.listItem()
+        .title('Content')
+        .child(
+          S.list()
+            .title('Content')
+            .items([
+              S.documentTypeListItem('blogPost').title('Blog Posts'),
+              S.documentTypeListItem('experience').title('Experience'),
+              S.documentTypeListItem('skillCategory').title('Skills'),
+            ]),
+        ),
+      S.divider(),
+      // Singleton — links directly to the single document, no list view
+      S.listItem()
+        .title('Site Settings')
+        .id('siteSettings')
+        .child(S.document().schemaType('siteSettings').documentId('siteSettings')),
+    ])
 
 export default defineConfig({
   name: 'default',
@@ -10,9 +34,15 @@ export default defineConfig({
   projectId: '9kzufcna',
   dataset: 'production',
 
-  plugins: [structureTool(), visionTool()],
+  plugins: [
+    structureTool({structure}),
+    visionTool(),
+  ],
 
   schema: {
     types: schemaTypes,
+    // Hide singleton types from the global "New document" menu
+    templates: (templates) =>
+      templates.filter(({schemaType}) => !singletonTypes.has(schemaType)),
   },
 })
