+++ studio/schemaTypes/documents/skillCategory.ts
