+++ studio/schemaTypes/index.ts
@@ -1,6 +1,8 @@
-import {blogPost} from './blogPost'
-import {experience} from './experience'
-import {skillCategory} from './skillCategory'
-import {siteSettings} from './siteSettings'
+import {blogPost} from './documents/blogPost'
+import {experience} from './documents/experience'
+import {skillCategory} from './documents/skillCategory'
+import {siteSettings} from './singletons/siteSettings'
 
 export const schemaTypes = [blogPost, experience, skillCategory, siteSettings]
+
+export const singletonTypes = new Set(['siteSettings'])
