+++ studio/schemaTypes/singletons/siteSettings.ts
@@ -4,8 +4,6 @@ export const siteSettings = defineType({
   name: 'siteSettings',
   title: 'Site Settings',
   type: 'document',
-  // Singleton — only one document of this type should exist
-  __experimental_actions: ['update', 'publish'],
   fields: [
     defineField({name: 'name', type: 'string', validation: (r) => r.required()}),
     defineField({
